[test] Add tests for game
What?: ref issue #7

Breaking Changes?
